pyproject.toml
[project]
name = "microsoftagentframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core framework + agents
    "agent-framework",
    "agent-framework-core",
    "azure-ai-projects",

    # Web stack
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",

    # HTTP clients / streaming
    "httpx[socks,http2]>=0.28,<0.29",
    "httpx-sse>=0.4.1",
    "requests",
    "aiohttp",

    # Azure SDKs
    "azure-identity",
    "azure-search-documents",

    # Data / DB / integration
    "psycopg[binary]>=3.2",
    "psycopg_pool>=3.2",
    "dapr-ext-fastapi",
    "isodate",

    # MCP + tools
    "fastmcp",
    "mcp[cli]",

    # Misc
    "python-dotenv",
    "openai",
]

# Optional: dev-only tools for local notebooks, etc.
[dependency-groups]
dev = [
    "ipykernel",
]

[tool.uv]
# Safer default: avoid prereleases unless explicitly needed.
prerelease = "allow"
